updates to screens. Removed second compose screen. Made a loop to test iNavigate
