+++ sampleapp/src/androidTest/java/com/progressive/sampleapp/screens/espresso/FinalScreen.kt
@@ -1,19 +1,28 @@
 package com.progressive.sampleapp.screens.espresso
 
 import android.app.Activity
+import androidx.compose.ui.test.junit4.ComposeTestRule
 import androidx.test.core.app.ActivityScenario
 import com.progressive.kherkin.sampleapp.R
 import com.progressive.kherkin.common.screen.Screen
 import com.progressive.kherkin.common.screen.Trait
+import com.progressive.kherkin.common.testcore.And
+import com.progressive.kherkin.common.testcore.Then
 import com.progressive.kherkin.espresso.steps.actions.ITouchButton
 import com.progressive.kherkin.espresso.testcore.Navigable
 import com.progressive.kherkin.espresso.testcore.PathSegment
 import com.progressive.kherkin.common.testcore.When
+import com.progressive.kherkin.compose.steps.actions.ITouchText
+import com.progressive.kherkin.compose.steps.actions.IWaitToSeeScreen
+import com.progressive.kherkin.compose.steps.testcore.ComposeNavigable
+import com.progressive.kherkin.compose.steps.testcore.ComposePathSegment
+import com.progressive.kherkin.espresso.steps.actions.IWaitToSeeScreen
 import com.progressive.sampleapp.activities.xml.FinalActivity
+import com.progressive.sampleapp.screens.compose.BasicComposeScreen
 import com.progressive.sampleapp.setup.SamplePreconditionsData
 import com.progressive.sampleapp.viewmodels.Destinations
 
-class FinalScreen : Screen(), Navigable {
+class FinalScreen : Screen(), Navigable, ComposeNavigable {
 
     override lateinit var activityScenario: ActivityScenario<FinalActivity>
     override val trait: Trait = Trait(R.id.buttonFinal)
@@ -43,6 +52,18 @@ class FinalScreen : Screen(), Navigable {
             )
         )
 
+        return pathSegments
+    }
+
+    override fun pathsToScreen(composeTestRule: ComposeTestRule): List<ComposePathSegment> {
+        val pathSegments = mutableListOf<ComposePathSegment>()
+        pathSegments.add(
+            ComposePathSegment(start = BasicComposeScreen(),
+                end = this,
+                step = {
+                    And.ITouchText("Navigate", composeTestRule)
+                })
+        )
         return pathSegments
     }
 }
